+++ frontend/rust-lib/flowy-sidecar/src/lib.rs
@@ -3,4 +3,5 @@ pub mod manager;
 pub mod parser;
 pub mod plugin;
 mod rpc_loop;
+mod rpc_object;
 mod rpc_peer;
